refactor: 패키지 구조 변경 - domain, repository 패키지 추가
