+++ src/test/java/repository/ReservationMemoryRepositoryTest.java
@@ -1,5 +1,8 @@
-import nextstep.Reservation;
-import nextstep.Theme;
+package repository;
+
+import nextstep.domain.Reservation;
+import nextstep.domain.Theme;
+import nextstep.repository.ReservationMemoryRepository;
 import nextstep.repository.ReservationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
